+++ lib/smart_core/operation/exceptions.rb
@@ -6,6 +6,26 @@ class SmartCore::Operation
   Error = Class.new(StandardError)
 
   # @api public
+  # @since 0.3.0
+  FatalError = Class.new(Error) do
+    # @return [SmartCore::Operation::Fatal]
+    #
+    # @api private
+    # @since 0.3.0
+    attr_reader :__operation_result__
+
+    # @param operation_result [SmartCore::Operation::Fatal]
+    # @return [void]
+    #
+    # @api private
+    # @since 0.3.0
+    def initialize(operation_result)
+      @__operation_result__ = operation_result
+      super()
+    end
+  end
+
+  # @api public
   # @since 0.2.0
   ArgumentError = Class.new(::ArgumentError)